+++ data/handleDB.php
@@ -485,6 +485,258 @@ class DbInitCommand extends Command
 }
 
 /**
+ * Build Extension draft JSON data
+ */
+class DbBuildExtCommand extends Command
+{
+    protected function configure()
+    {
+        $this->setName('db:build:ext')
+            ->setDescription('Build Extension draft JSON data for SQLite compatinfo database')
+            ->addArgument(
+                'extension',
+                InputArgument::REQUIRED,
+                'extension to extract components (case insensitive)'
+            )
+            ->addArgument(
+                'ext_min',
+                InputArgument::OPTIONAL,
+                'extension MIN version',
+                '0.1.0'
+            )
+            ->addArgument(
+                'php_min',
+                InputArgument::OPTIONAL,
+                'php MIN version',
+                '5.3.0'
+            )
+        ;
+    }
+
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $extension = trim($input->getArgument('extension'));
+        $extMin    = trim($input->getArgument('ext_min'));
+        $phpMin    = trim($input->getArgument('php_min'));
+
+        $pdo = Environment::initRefDb();
+
+        $stmtExtension = $pdo->prepare(
+            'SELECT e.id as "ext_name_fk", e.name as "name"' .
+            ' FROM bartlett_compatinfo_extensions e' .
+            ' WHERE e.name = :name COLLATE NOCASE'
+        );
+
+        $inputParameters = array(':name' => $extension);
+        $stmtExtension->execute($inputParameters);
+        $result = $stmtExtension->fetch(PDO::FETCH_ASSOC);
+
+        if (!$result) {
+            // unknown extension
+            $extId   = -1;
+            $extName = $extension;
+        } else {
+            $extId   = $result['ext_name_fk'];
+            $extName = $result['name'];
+        }
+
+        // EXTENSIONS
+        $data = ['id' => $extId, 'name' => $extName];
+        $output->writeln(
+            sprintf(
+                '<info>%s.extensions.json</info>',
+                ucfirst(strtolower($extension))
+            )
+        );
+        $output->writeln(
+            json_encode($data, JSON_PRETTY_PRINT)
+        );
+
+        $re = new \ReflectionExtension($extension);
+        $classes = $re->getClassNames();
+        natsort($classes);
+
+        if (is_array($classes)) {
+            // CLASSES
+            $output->writeln(
+                sprintf(
+                    '<info>%s.classes.json</info>',
+                    ucfirst(strtolower($extension))
+                )
+            );
+            $this->buildClasses($extId, $extMin, $phpMin, $re, $output);
+
+            // INTERFACES
+            $output->writeln(
+                sprintf(
+                    '<info>%s.interfaces.json</info>',
+                    ucfirst(strtolower($extension))
+                )
+            );
+            $this->buildInterfaces($extId, $extMin, $phpMin, $re, $output);
+
+            // METHODS
+            $output->writeln(
+                sprintf(
+                    '<info>%s.methods.json</info>',
+                    ucfirst(strtolower($extension))
+                )
+            );
+            $this->buildMethods($extId, $extMin, $phpMin, $classes, $output);
+
+            // CLASSES CONSTANTS
+            $output->writeln(
+                sprintf(
+                    '<info>%s.const.json</info>',
+                    ucfirst(strtolower($extension))
+                )
+            );
+            $this->buildClassConstants($extId, $extMin, $phpMin, $classes, $output);
+        }
+    }
+
+    private function buildClasses($extId, $extMin, $phpMin, \ReflectionExtension $re, OutputInterface $output)
+    {
+        $objects = [];
+
+        foreach ($re->getClasses() as $className => $rc) {
+            if ($rc->isInterface()) {
+                continue;
+            }
+
+            $data = [
+                'ext_name_fk' => $extId,
+                'name' => $className,
+                'ext_min' => $extMin,
+                'ext_max' => '',
+                'php_min' => $phpMin,
+                'php_max' => '',
+            ];
+            $objects[$className] = $data;
+        }
+
+        ksort($objects, SORT_NATURAL);
+
+        $output->writeln(
+            json_encode(array_values($objects), JSON_PRETTY_PRINT)
+        );
+    }
+
+    private function buildInterfaces($extId, $extMin, $phpMin, \ReflectionExtension $re, OutputInterface $output)
+    {
+        $objects = [];
+
+        foreach ($re->getClasses() as $className => $rc) {
+            if (!$rc->isInterface()) {
+                continue;
+            }
+
+            $data = [
+                'ext_name_fk' => $extId,
+                'name' => $className,
+                'ext_min' => $extMin,
+                'ext_max' => '',
+                'php_min' => $phpMin,
+                'php_max' => '',
+            ];
+            $objects[$className] = $data;
+        }
+
+        ksort($objects, SORT_NATURAL);
+
+        $output->writeln(
+            json_encode(array_values($objects), JSON_PRETTY_PRINT)
+        );
+    }
+
+    private function buildMethods($extId, $extMin, $phpMin, array $classes, OutputInterface $output)
+    {
+        $methods = [];
+
+        foreach ($classes as $className) {
+            $rc = new \ReflectionClass($className);
+
+            $meth = $rc->getMethods(\ReflectionMethod::IS_PUBLIC);
+            natsort($meth);
+
+            foreach ($meth as $method) {
+                $methodName = $method->getName();
+
+                try {
+                    $method->getPrototype();
+                    // if there is one method prototype, don't check it
+                    if ($output->isVerbose()) {
+                        $output->writeln(
+                            sprintf('<comment>%s</comment> is a prototype', $methodName)
+                        );
+                    }
+                    continue;
+                } catch (\ReflectionException $e) {
+                }
+
+                $from = $method->getDeclaringClass()->getName();
+
+                if ($from !== $className) {
+                    if ($output->isVerbose()) {
+                        $output->writeln(
+                            sprintf('<comment>%s</comment> inherit from <comment>%s</comment>', $methodName, $from)
+                        );
+                    }
+                    continue;
+                }
+
+                $static = $method->isStatic() ? 'true' : 'false';
+
+                $data = [
+                    'ext_name_fk' => $extId,
+                    'class_name' => $className,
+                    'name' => $methodName,
+                    'static' => $static,
+                    'ext_min' => $extMin,
+                    'ext_max' => '',
+                    'php_min' => $phpMin,
+                    'php_max' => '',
+                ];
+                $methods[] = $data;
+            }
+        }
+
+        $output->writeln(
+            json_encode($methods, JSON_PRETTY_PRINT)
+        );
+    }
+
+    private function buildClassConstants($extId, $extMin, $phpMin, array $classes, OutputInterface $output)
+    {
+        $constants = [];
+
+        foreach ($classes as $className) {
+            $rc = new \ReflectionClass($className);
+
+            $const = array_keys($rc->getConstants());
+            natsort($const);
+
+            foreach ($const as $constName) {
+                $data = [
+                    'ext_name_fk' => $extId,
+                    'class_name' => $className,
+                    'name' => $constName,
+                    'ext_min' => $extMin,
+                    'ext_max' => '',
+                    'php_min' => $phpMin,
+                    'php_max' => '',
+                ];
+                $constants[] = $data;
+            }
+        }
+
+        $output->writeln(
+            json_encode($constants, JSON_PRETTY_PRINT)
+        );
+    }
+}
+
+/**
  * Update JSON files when a new PHP version is released.
  */
 class DbReleaseCommand extends Command
@@ -870,6 +1122,7 @@ class DbHandleApplication extends Application
 
         $defaultCommands[] = new DbBackupCommand();
         $defaultCommands[] = new DbInitCommand();
+        $defaultCommands[] = new DbBuildExtCommand();
         $defaultCommands[] = new DbReleaseCommand();
 
         return $defaultCommands;
@@ -907,5 +1160,5 @@ class DbHandleApplication extends Application
     }
 }
 
-$application = new DbHandleApplication('Database handler for CompatInfo', '1.22.0');
+$application = new DbHandleApplication('Database handler for CompatInfo', '1.23.0');
 $application->run();